requirements.txt
streamlit
julep
requests
python-dotenv
requests
